forgot

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -9,6 +9,7 @@ use App\Models\Category;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Http\Resources\Json\JsonResource;
+use Illuminate\Support\Facades\Auth;
 
 class CategoryController extends Controller
 {
@@ -20,18 +21,23 @@ class CategoryController extends Controller
         $this->categoryService = $categoryService;
     }
     public function index(){
-        
-        return view('admin.categories.index', [
-            'title' => 'Danh mục sản phẩm',
-            'list_category' => Category::orderByDesc('id')->search()->paginate(30)
-        ]);
+        if(Auth::user()->user_type == 'AD'){
+            return view('admin.categories.index', [
+                'title' => 'Danh mục sản phẩm',
+                'list_category' => Category::orderByDesc('id')->search()->paginate(30)
+            ]);
+        }
+        return redirect()->back();        
     }
 
     public function create(){
-        return view('admin.categories.create', [
-            'title' => 'Thêm danh mục ',
-            'parentCategories' => $this->categoryService->getParent()  
-        ]);
+        if(Auth::user()->user_type == 'AD'){
+            return view('admin.categories.create', [
+                'title' => 'Thêm danh mục ',
+                'parentCategories' => $this->categoryService->getParent()  
+            ]);
+        }
+        return redirect()->back();
     }
 
     public function store(CreateFormRequest $request){
@@ -42,11 +48,14 @@ class CategoryController extends Controller
 
     public function show(Category $category)
     {   
-        return view('admin.categories.edit',[
-            'title' => 'Chỉnh sửa danh mục: ' . $category->name,
-            'category' => $category,
-            'parentCategories' => $this->categoryService->getParent()  
-        ]);
+        if(Auth::user()->user_type == 'AD'){
+            return view('admin.categories.edit',[
+                'title' => 'Chỉnh sửa danh mục: ' . $category->name,
+                'category' => $category,
+                'parentCategories' => $this->categoryService->getParent()  
+            ]);        
+        }
+        return redirect()->back();
 
     }
 
@@ -58,15 +67,19 @@ class CategoryController extends Controller
 
     public function destroy(Request $request):JsonResponse
     {
-        $result = $this->categoryService->destroy($request);
-        if($result){
-            return $request->json([
-                'error' => false,
-                'message' => 'Xóa danh mục thành công.'
+        if(Auth::user()->user_type == 'AD'){
+            $result = $this->categoryService->destroy($request);
+            if($result){
+                return $request->json([
+                    'error' => false,
+                    'message' => 'Xóa danh mục thành công.'
+                ]);
+            }
+            return  response()->json([
+                'error' => true
             ]);
-        }
-        return  response()->json([
-            'error' => true
-        ]);
+        }        
+        return redirect()->back();
+
     }
 }
